+++ tests/tests/models/test_collection.py
@@ -256,6 +256,32 @@ class TestCollection:
             assert next(storage.image_collection.detect_item_type(file)) == 'image'
 
 
+class TestDeleteCollection:
+    """
+    Тестирование удаления галереи, включающей элементы с вариациями.
+    """
+    @staticmethod
+    def init_class(storage):
+        storage.collection = CompleteCollection.objects.create()
+
+        file_item = FileItem()
+        file_item.attach_to(storage.collection)
+        with open(CALLIPHORA_FILEPATH, 'rb') as fp:
+            file_item.attach_file(fp, name='file_c3.jpg')
+        file_item.save()
+
+        image_item = ImageItem()
+        image_item.attach_to(storage.collection)
+        with open(NASA_FILEPATH, 'rb') as fp:
+            image_item.attach_file(fp, name='image_c3.jpg')
+        image_item.save()
+
+        yield
+
+    def test_deletion(self, storage):
+        storage.collection.delete()
+
+
 class CollectionItemMixin:
     owner_app_label = ''
     owner_model_name = ''