+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.6.3](https://github.com/dldevinc/paper-uploads/tree/v0.6.3) - 2021-11-03
 ### Bug Fixes
 - Fixed image recut with `python-rq`.
+- Fixed `RecursionError` when deleting a collection.
 
 ## [0.6.2](https://github.com/dldevinc/paper-uploads/tree/v0.6.2) - 2021-11-02
 ### Features